src/requirements.txt
fastapi==0.124.4
uvicorn[standard]==0.30.0
python-multipart==0.0.12
python-dotenv==1.0.1
pydantic-settings==2.6.0
aiofiles==24.1.0
PyMuPDF==1.26.7
asyncpg==0.30.0
alembic == 1.17.2
psycopg2==2.9.11
pgvector==0.4.2
google-generativeai==0.8.6
tqdm==4.67.1
pandas==2.3.3
ragas==0.4.3
datasets==4.4.2
